requirements.txt
blinker==1.7.0
bluedot==2.0.0
click==8.1.7
colorzero==2.0
Flask==3.0.2
gpiozero==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
RPi.GPIO==0.7.1
Werkzeug==3.0.1
